refactor(ReceiverSection): 정적 상수들 변수화
