+++ src/pages/OrderPage/constants/table.ts
@@ -1,0 +1,5 @@
+export const TABLE_CELL_CONSTANTS = {
+  NAME: "이름",
+  PHONE: "전화번호",
+  QUANTITY: "수량",
+} as const;
